requirements.txt
pyscard>=2.0.0
pycryptodome>=3.15.0